requirements.txt
aenum==3.1.16
aiohappyeyeballs==2.6.1
aiohttp==3.12.11
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
attrs==25.3.0
billiard==4.2.1
blinker==1.9.0
celery==5.5.3
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
Deprecated==1.2.18
dnspython==2.7.0
et_xmlfile==2.0.0
eventlet==0.40.2
filelock==3.18.0
Flask==3.1.1
flask-cors==6.0.1
fonttools==4.58.2
frozenlist==1.6.2 
fsspec==2025.5.1
future==1.0.0
greenlet==3.2.3
gunicorn==23.0.0
huggingface-hub==0.34.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
kombu==5.5.4
line-bot-sdk==3.17.1
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
multidict==6.4.4
networkx==3.5
numpy==2.2.6
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
packaging==25.0
pandas==2.3.0
pillow==11.2.1
prompt_toolkit==3.0.51
propcache==0.3.1
psutil==7.0.0
py-cpuinfo==9.0.0
pydantic==2.11.5
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
redis==6.3.0
regex==2025.7.34
requests==2.32.3
safetensors==0.6.1
scipy==1.15.3
setuptools==80.9.0
six==1.17.0
sympy==1.14.0
tokenizers==0.21.4
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
transformers==4.55.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
ultralytics==8.3.152
ultralytics-thop==2.0.14
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.17.2
yarl==1.20.0
